+++ predicators/agent_sdk/agent_session_mixin.py
@@ -128,12 +128,16 @@ class AgentSessionMixin:
                 tools=tools,
             )
 
+            extra_names = [
+                getattr(t, "name", "") for t in
+                self._tool_context.extra_mcp_tools]
             self._agent_session = AgentSessionManager(
                 system_prompt=self._get_agent_system_prompt(),
                 mcp_server=mcp_server,
                 log_dir=self._get_log_dir(),
                 model_name=CFG.agent_sdk_model_name,
-                allowed_tools=get_allowed_tool_list(tool_names),
+                allowed_tools=get_allowed_tool_list(
+                    tool_names, extra_names=extra_names or None),
             )
 
         if self._agent_session_id is not None:
@@ -186,11 +190,12 @@ class AgentSessionMixin:
         self,
         predicates: Set[Predicate],
         options: Set[ParameterizedOption],
+        name: str = "agent_plan",
     ) -> BaseExplorer:
         """Create an agent explorer with tool_context and agent_session."""
         self._ensure_agent_session()
         return create_explorer(
-            "agent_plan",
+            name,
             predicates,
             options,
             self._types,  # type: ignore[attr-defined]
